requirements.txt
crewai
crewai-tools
streamlit
python-dotenv
langchain-google-genai